Improve uninstaller creation
Instead of calling the installer script recursivly to create the
uninstaller, do it properly. Now we don't embed uninstaller for other
architecture than what we are installing. Tracking the install mode for the
uninstaller is solved by adding a commandline option in the
UninstallString, and parsing for this when uninstalling. If not present,
the user mode used to launch the uninstaller will be used to determine
which mode should be used.