Fix permission problem
By default WinPython sets the home env variable to the "settings" folder in the
WinPython folder. However, non-admin users don't have write access to this
folder, which means that any program to writes to that folder fails.

This fixes the issue by providing scripts to start spyder, qtconsole and notebook without modifying the default HOME env value.

This also removes the matplotlibrc file that is not longer necessary since
we're not going to release any py2 version of the bundle.